[FIX] mail_autosusbscribe: Removing external dependencies and adding context to tests
The external dependency odoo_test_helper must be defined in test-requirements.txt
not in the manifest.
Also, a context has been added to the tests to avoid executing the code of the
"_message_get_default_recipients" method added by this module in a test
environment as long as it is not the module itself that is being tested.

TT48102